requirements.txt
streamlit>=1.30.0
requests>=2.28.0
python-dotenv>=0.19.0
pytz>=2021.1